+++ src/components/feedback/Toast/index.ts
@@ -1,7 +1,7 @@
 /**
  * Toast 모듈 내보내기
  */
-export { ToastProvider, useToast } from './ToastContext';
+export { ToastProvider } from './ToastContext';
+export { useToast } from './useToast';
 export { Toast, ToastContainer } from './Toast';
-export type { ToastItem, ToastType } from './ToastContext';
-
+export type { ToastItem, ToastType, ToastContextType } from './ToastTypes';
