+++ src/hooks/useAutoSave.ts
@@ -50,7 +50,7 @@ import { debounce } from '@/common/utils';
  * @param {any} data - 저장할 데이터
  * @param {number} delay - Debounce 지연 시간 (기본 1000ms)
  */
-export const useAutoSave = (data: any, delay: number = 1000) => {
+export const useAutoSave = (data: unknown, delay: number = 1000) => {
   const { setSaveStatus } = useProjectStore();
   const previousDataRef = useRef<string | undefined>(undefined);
